% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_builder.R
\name{build_seed}
\alias{build_seed}
\title{Provides a function to initialize from prototypes or from clusters a penalized mixtures of von Mises-Fisher.}
\usage{
build_seed(X, nb_run = 10, nb_init = 10, ks = 2:6)
}
\arguments{
\item{X}{Data : matrix with column as features.}

\item{nb_run}{Number of runs.}

\item{nb_init}{Number of initialisations.}

\item{ks}{Number of cluster to be tested.}
}
\value{
returns a list of seed.
}
\description{
Provides a function to initialize from prototypes or from clusters a penalized mixtures of von Mises-Fisher.
}
