% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_following_vmf.R
\name{movMF_beta_path}
\alias{movMF_beta_path}
\title{Provides a function  the trade-off between the sparsity term and the likelihood one with a simple path following algorithm.}
\usage{
movMF_beta_path(
  X,
  K,
  Theta = NULL,
  shared_kappa = FALSE,
  restart = NULL,
  nb_beta = NULL,
  maxiter = 1000,
  maxiterfp = 10,
  prec = sqrt(.Machine$double.eps),
  kappamax = 1e+06,
  mode = "random",
  sk_runs = 1,
  interpolate = 1,
  min_rel_inc = 0,
  save_tau = FALSE,
  hard_assign = TRUE,
  save_path = c("full", "statistics", "nothing"),
  statistics = movMF_extract_stats,
  verbose = 0
)
}
\arguments{
\item{X}{Data : matrix with column as features.}

\item{K}{The number of mixture components (or clusters).}

\item{Theta}{Theta is the initial parameter list.}

\item{shared_kappa}{If all components share kappa or not.}

\item{restart}{Number of restart.}

\item{nb_beta}{number of steps of the path following approach.}

\item{maxiter}{The maximum number of iteration.}

\item{maxiterfp}{The maximum number of iteration for the fix poit strategy.}

\item{prec}{Precision.}

\item{kappamax}{Kappa max value.}

\item{mode}{Method to initialize: "random", "spread", "skmeans".}

\item{sk_runs}{Number of runs for the initialization for skmeans.}

\item{interpolate}{interpolate paramters.}

\item{min_rel_inc}{Minimum increment for beta.}

\item{save_tau}{Save responsibilities Tau.}

\item{hard_assign}{Hard assignment.}

\item{save_path}{save information about the path with 3 options "full", "statistics", "nothing".}

\item{statistics}{by default internal function movMF_extract_stats to extract ARI and NMI will be used.}

\item{verbose}{Verbose.}

\item{beta}{Penalty parameter.}
}
\value{
returns two lists depending on the statistic parameter chosen:
\itemize{
\item First list is composed by the best models along the path selected depending on the criterion.
\item Second list is a list of models or their principal statistics or only the dense model.
}
}
\description{
Provides a function  the trade-off between the sparsity term and the likelihood one with a simple path following algorithm.
}
