% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_vmf.R
\name{movMF_EM}
\alias{movMF_EM}
\title{Provides a function to build and fit a penalized mixtures of von Mises-Fisher.}
\usage{
movMF_EM(
  X,
  K,
  beta,
  Theta = NULL,
  shared_kappa = FALSE,
  restart = NULL,
  maxiter = 1000,
  maxiterfp = 10,
  prec = sqrt(.Machine$double.eps),
  kappamax = 1e+06,
  mode = "random",
  sk_runs = 1,
  save_tau = TRUE,
  hard_assign = TRUE,
  verbose = 0
)
}
\arguments{
\item{X}{Data : matrix with column as features}

\item{K}{The number of mixture components (or clusters).}

\item{beta}{Penalty parameter.}

\item{Theta}{Theta is the initial parameter list.}

\item{shared_kappa}{If all components share kappa or not.}

\item{restart}{Number of restart.}

\item{maxiter}{The maximum number of iteration.}

\item{maxiterfp}{The maximum number of iteration for the fix poit strategy.}

\item{prec}{Precision.}

\item{kappamax}{Kappa max value.}

\item{mode}{Method to initialize: "random", "spread", "skmeans".}

\item{sk_runs}{Number of runs for the initialization for skmeans.}

\item{save_tau}{Save tau.}

\item{hard_assign}{Hard assignment.}

\item{verbose}{Verbose.}
}
\value{
returns a list including the following attributes:
\itemize{
\item theta: mu times kappa.
\item kappa : kappa
\item L: log likelihood.
\item alpha: a vector of cluster proportions.
\item iter: the total number of iteration needed for convergence.
\item Prob: matrix of probability to be in a cluster
}
}
\description{
Provides a function to build and fit a penalized mixtures of von Mises-Fisher.
}
