% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movmf_init.R
\name{movMF_initialisation_internal}
\alias{movMF_initialisation_internal}
\title{Provides a function to initialize from prototypes or from clusters a penalized mixtures of von Mises-Fisher.}
\usage{
movMF_initialisation_internal(X, proto = NULL, cluster = NULL, shared_kappa)
}
\arguments{
\item{X}{Data : matrix with column as features.}

\item{proto}{Directional means.}

\item{cluster}{Initial clusters.}

\item{shared_kappa}{If all components share kappa or not.}
}
\value{
returns a list including the following attributes:
\itemize{
\item mu: directional means.
\item alpha: mixture proportions.
\item kappa: concentration parameters.
}
}
\description{
Provides a function to initialize from prototypes or from clusters a penalized mixtures of von Mises-Fisher.
}
