% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Provides a function to plot the data set according to the result of a mixture of von Mises-Fisher distribution.}
\usage{
plot_data(movMFresult, X, with_color = FALSE)
}
\arguments{
\item{X}{dataset used by the function movMF_EM.}

\item{with_color}{add colors to the plot.}

\item{movMFresul}{results from the function movMF_EM.}
}
\value{
returns a plot of the dataset.
}
\description{
Provides a function to plot the data set according to the result of a mixture of von Mises-Fisher distribution.
}
