% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prototypes.R
\name{plot_proto}
\alias{plot_proto}
\title{Provides a function to plot the directional means of a mixture of von Mises-Fisher distribution.}
\usage{
plot_proto(movMFresult, with_color = FALSE)
}
\arguments{
\item{with_color}{add colors to the plot.}

\item{movMFresul}{results from the function movMF_EM.}
}
\value{
returns a plot of the directional means.
}
\description{
Provides a function to plot the directional means of a mixture of von Mises-Fisher distribution.
}
