% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_vmf.R
\name{movMF_EM}
\alias{movMF_EM}
\title{Provides a function to build and fit a penalized mixtures of von Mises-Fisher.}
\usage{
movMF_EM(
  X,
  K,
  beta,
  Theta = NULL,
  shared_kappa = FALSE,
  restart = NULL,
  maxiter = 1000,
  maxiterfp = 10,
  prec = sqrt(.Machine$double.eps),
  kappamax = 1e+06,
  mode = "random",
  sk_runs = 1,
  save_tau = TRUE,
  hard_assign = TRUE,
  verbose = 0
)
}
\arguments{
\item{X}{Data : matrix with column as features}

\item{K}{The number of mixture components (or clusters).}

\item{beta}{Penalty parameter.}

\item{Theta}{Theta is the initial parameter list.}

\item{shared_kappa}{If all components share kappa or not.}

\item{restart}{Number of restart.}

\item{maxiter}{The maximum number of iteration.}

\item{maxiterfp}{The maximum number of iteration for the fix poit strategy.}

\item{prec}{Precision.}

\item{kappamax}{Kappa max value.}

\item{mode}{Method to initialize: "random", "spread", "skmeans".}

\item{sk_runs}{Number of runs for the initialization for skmeans.}

\item{save_tau}{Save responsibilities Tau.}

\item{hard_assign}{Hard assignment.}

\item{verbose}{Verbose.}
}
\value{
returns a list including the following attributes:
\itemize{
\item mu: directional means.
\item alpha: mixture proportions.
\item kappa: concentration parameters.
\item Tau: responsibilities.
\item shared_kappa: TRUE if the model shares kappa.
\item beta: penalty parameter.
\item logL: Log-likelihood.
\item fpiter: fix point iterations.
\item iter: iterations.
\item logLikelihood: best LogLikelihood.
\item penLogLikelihood: best penalized LogLikelihood.
\item completeLogLikelihood: best complete LogLikelihood.
\item status:
\itemize{
\item 0: converged,
\item 1: maxiter reached,
\item 2: not converging to a K cluster solution,
\item 3: infinite variance.
}
\item n: number of rows of X.
\item cluster: hard assignment.
\item converged: nb of status converged.
\item infinite: nb of status infinite.
\item empty: nb of status not converging to a K cluster solution.
\item itermax: iter maximum.
\item total_iterations: nb total iteration.
\item total_internal_iterations: nb total iteration for fix point strategy.
\item all_logLikelihood: all log Likelihood.
\item all_penLogLikelihood: all penalized log Likelihood.
\item IC: List of Information criteria BIC, AIC, RIC, RICc, EBIC with :
\itemize{
\item d: we keep all parameters,
\item z: we count non zero values per directional mean and keep at least one parameter,
\item v: we count the variables with at least one non zero coordinate on a directional mean.
}
\item sparsity: sparsity of mu.
}
}
\description{
Provides a function to build and fit a penalized mixtures of von Mises-Fisher.
}
